Close #329 - e2e functions generator
When type `npx nestia e2e` command, `@nestia/sdk` will analyze NestJS server code, and generates e2e test functions for every API functions provided by the NestJS backend server.

The automatically generated e2e test function will based on `typia.random<T>()` function. Therefore, 400 error never be occured, but such randomly generated e2e test codes can occure domain logic error. So, I recommend you to change those automatically generated e2e functions code, to be suitable for your domain logics.

+++ packages/sdk/src/generates/SdkGenerator.ts
@@ -1,44 +1,42 @@
 import fs from "fs";
 import NodePath from "path";
-import ts from "typescript";
 
 import { INestiaConfig } from "../INestiaConfig";
 import { IRoute } from "../structures/IRoute";
-import { FileGenerator } from "./FileGenerator";
+import { SdkFileProgrammer } from "./internal/SdkFileProgrammer";
 
 export namespace SdkGenerator {
-    export async function generate(
-        _checker: ts.TypeChecker,
-        config: INestiaConfig,
-        routeList: IRoute[],
-    ): Promise<void> {
-        // PREPARE NEW DIRECTORIES
-        try {
-            await fs.promises.mkdir(config.output!);
-        } catch {}
+    export const generate =
+        (config: INestiaConfig) =>
+        async (routes: IRoute[]): Promise<void> => {
+            // PREPARE NEW DIRECTORIES
+            try {
+                await fs.promises.mkdir(config.output!);
+            } catch {}
 
-        // BUNDLING
-        const bundle: string[] = await fs.promises.readdir(BUNDLE_PATH);
-        for (const file of bundle) {
-            const current: string = `${BUNDLE_PATH}/${file}`;
-            const stats: fs.Stats = await fs.promises.stat(current);
+            // BUNDLING
+            const bundle: string[] = await fs.promises.readdir(BUNDLE_PATH);
+            for (const file of bundle) {
+                const current: string = `${BUNDLE_PATH}/${file}`;
+                const stats: fs.Stats = await fs.promises.stat(current);
 
-            if (stats.isFile() === true) {
-                const content: string = await fs.promises.readFile(
-                    current,
-                    "utf8",
-                );
-                await fs.promises.writeFile(
-                    `${config.output}/${file}`,
-                    content,
-                    "utf8",
-                );
+                if (stats.isFile() === true) {
+                    const content: string = await fs.promises.readFile(
+                        current,
+                        "utf8",
+                    );
+                    if (fs.existsSync(`${config.output}/${file}`) === false)
+                        await fs.promises.writeFile(
+                            `${config.output}/${file}`,
+                            content,
+                            "utf8",
+                        );
+                }
             }
-        }
 
-        // FUNCTIONAL
-        await FileGenerator.generate(config, routeList);
-    }
+            // FUNCTIONAL
+            await SdkFileProgrammer.generate(config)(routes);
+        };
 
     export const BUNDLE_PATH = NodePath.join(
         __dirname,
@@ -46,5 +44,6 @@ export namespace SdkGenerator {
         "..",
         "assets",
         "bundle",
+        "api",
     );
 }
